[Cb-25] Create a parent component to display stock items (#15)
* Added a new parent component to display the data

* Changed a class component to a fucntional one

* Removed test text

* Moved components around so that they are displayed in the stock details component

* Removed proxy

* Removed unused comments

* Removed unused component

* Removed route from this branch

* Moved small issues to other pr

* Renamed cloud picture
