+++ frontend/src/components/StockDetails/StockQuote/StockQuote.css
